Added Movie to the domain

+++ bookshelf-pro-domain/src/main/java/org/a1cey/bookshelf_pro_domain/bookshelf/bookshelf_entry/consumption/MediaItemConsumptionProgress.java
@@ -12,7 +12,7 @@ public interface MediaItemConsumptionProgress {
 
     /**
      * Create a new MediaItemConsumptionProgress that is valid for the same media item.
-     * Provide the new current consumption value.
+     * Provide the new current consumption time.
      */
-    MediaItemConsumptionProgress update(int newCurrent);
+    MediaItemConsumptionProgress update(long newCurrent);
 }
